+++ R/cnefe_counts.R
@@ -166,6 +166,7 @@ cnefe_counts <- function(
     index = cnefe_index,
     cache = cache,
     cache_dir = cache_dir,
+    year = year,
     verbose = verbose,
     retry_timeouts = c(300L, 600L, 1800L)
   )
@@ -387,6 +388,7 @@ cnefe_counts <- function(
     index = cnefe_index,
     cache = cache,
     cache_dir = cache_dir,
+    year = year,
     verbose = verbose,
     retry_timeouts = c(300L, 600L, 1800L)
   )