+++ R/read_cnefe.R
@@ -120,6 +120,7 @@ read_cnefe <- function(
     index = cnefe_index,
     cache = cache,
     cache_dir = cache_dir,
+    year = year,
     verbose = verbose,
     retry_timeouts = c(300L, 600L, 1800L)
   )